Implement opt-in change detection in pull requests
This patch introduces a `pull-request-change-detection` action input
that allows the end-users to request running `ansible-test` in change
detection mode. It is only supported in for pull requests at the
moment because `ansible-test` does not currently know how to interact
with the environment variables that GitHub Actions CI/CD workflows
set [1].

With this feature, the users will be able to save some of the CI
compute time due to `ansible-test` skipping the tests that are not
affected by the changes in PRs.

[1] https://github.com/ansible-community/ansible-test-gh-action/pull/49/files#r1023412786