Implement toggles for integration tests flags
Currently, all of `--retry-on-error`, `--continue-on-error` and
`--diff` are always passed to `ansible-test` when this action is
running integration tests.

This patch converts each of them into individual inputs that are
enabled by default but can be set to `false` when it's useful to
the tested collection.

The new input names are:

* `integration-continue-on-error`
* `integration-diff`
* `integration-retry-on-error`